Hoàn thành

+++ src/Controllers/BackendController.php
@@ -2,16 +2,16 @@
 
 namespace Minhbang\Article\Controllers;
 
+use CategoryManager;
+use Datatables;
+use Minhbang\Article\Article;
 use Minhbang\Article\ArticleTransformer;
-use Minhbang\Kit\Extensions\BackendController as BaseController;
-use Minhbang\Kit\Traits\Controller\QuickUpdateActions;
 use Minhbang\Article\Request as ArticleRequest;
+use Minhbang\Kit\Extensions\BackendController as BaseController;
+use Minhbang\Kit\Extensions\DatatableBuilder as Builder;
+use Minhbang\Kit\Traits\Controller\QuickUpdateActions;
 use Request;
 use Session;
-use Minhbang\Article\Article;
-use CategoryManager;
-use Minhbang\Kit\Extensions\DatatableBuilder as Builder;
-use Datatables;
 use Status;
 
 /**
@@ -20,8 +20,10 @@ use Status;
  *
  * @package Minhbang\Article
  */
-class BackendController extends BaseController {
+class BackendController extends BaseController
+{
     use QuickUpdateActions;
+
     /**
      * @var \Minhbang\Category\Root
      */
@@ -30,9 +32,10 @@ class BackendController extends BaseController {
     /**
      * BackendController constructor.
      */
-    public function __construct() {
+    public function __construct()
+    {
         parent::__construct();
-        $this->categoryManager = CategoryManager::of( Article::class );
+        $this->categoryManager = CategoryManager::of(Article::class);
     }
 
     /**
@@ -40,8 +43,9 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function select() {
-        return response()->json( Article::forSelectize( Request::get( 'title' ) )->get()->all() );
+    public function select()
+    {
+        return response()->json(Article::forSelectize(Request::get('title'))->get()->all());
     }
 
     /**
@@ -49,87 +53,94 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\View\View
      */
-    public function index( Builder $builder ) {
+    public function index(Builder $builder)
+    {
         $typeName = $this->categoryManager->typeName();
         $this->buildHeading(
-            [ trans( 'common.manage' ), $typeName ],
+            [trans('common.manage'), $typeName],
             'fa-newspaper-o',
-            [ '#' => $typeName ]
+            ['#' => $typeName]
         );
-        $builder->ajax( route( 'backend.article.data' ) );
+        $builder->ajax(route('backend.article.data'));
 
-        $html = $builder->columns( [
-            [ 'data' => 'id', 'name' => 'id', 'title' => 'ID', 'class' => 'min-width text-center' ],
+        $html = $builder->columns([
+            ['data' => 'id', 'name' => 'id', 'title' => 'ID', 'class' => 'min-width text-center'],
             [
-                'data'  => 'title',
-                'name'  => 'title',
-                'title' => trans( 'article::common.title' ),
+                'data' => 'title',
+                'name' => 'title',
+                'title' => trans('article::common.title'),
             ],
-            [ 'data' => 'author', 'name' => 'author', 'title' => trans( 'article::common.user' ), 'class' => 'min-width', ],
             [
-                'data'  => 'updated_at',
-                'name'  => 'updated_at',
-                'title' => trans( 'common.updated_at' ),
+                'data' => 'author',
+                'name' => 'users.username',
+                'title' => trans('article::common.user'),
                 'class' => 'min-width',
             ],
             [
-                'data'       => 'status',
-                'name'       => 'status',
-                'title'      => trans( 'common.status' ),
-                'class'      => 'min-width',
-                'orderable'  => false,
+                'data' => 'updated_at',
+                'name' => 'updated_at',
+                'title' => trans('common.updated_at'),
+                'class' => 'min-width',
+            ],
+            [
+                'data' => 'status',
+                'name' => 'status',
+                'title' => trans('common.status'),
+                'class' => 'min-width',
+                'orderable' => false,
                 'searchable' => false,
             ],
-        ] )->addAction( [
-            'data'  => 'actions',
-            'name'  => 'actions',
-            'title' => trans( 'common.actions' ),
+        ])->addAction([
+            'data' => 'actions',
+            'name' => 'actions',
+            'title' => trans('common.actions'),
             'class' => 'min-width',
-        ] );
+        ]);
 
-        return view( 'article::backend.index', compact( 'typeName', 'html' ) );
+        return view('article::backend.index', compact('typeName', 'html'));
     }
-
 
     /**
      * Danh sách Article theo định dạng của Datatables.
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function data() {
-        // TODOL giải quyết mặc định order updated: ->orderUpdated()
-        $query = Article::queryDefault()->withAuthor()->orderUpdated()->categorized( $this->categoryManager->node() );
-        if ( Request::has( 'search_form' ) ) {
+    public function data()
+    {
+        $query =
+            Article::queryDefault()->ready('update')->withAuthor()->orderUpdated()->categorized($this->categoryManager->node());
+        if (Request::has('search_form')) {
             $query = $query
-                ->searchWhereBetween( 'articles.created_at', 'mb_date_vn2mysql' )
-                ->searchWhereBetween( 'articles.updated_at', 'mb_date_vn2mysql' );
+                ->searchWhereBetween('articles.created_at', 'mb_date_vn2mysql')
+                ->searchWhereBetween('articles.updated_at', 'mb_date_vn2mysql');
         }
 
-        return Datatables::of( $query )->setTransformer( new ArticleTransformer() )->make( true );
+        return Datatables::of($query)->setTransformer(new ArticleTransformer())->make(true);
     }
 
     /**
      * @return \Illuminate\View\View
      * @throws \Laracasts\Presenter\Exceptions\PresenterException
      */
-    public function create() {
+    public function create()
+    {
         $article = new Article();
-        $url = route( 'backend.article.store' );
+        $url = route('backend.article.store');
         $method = 'post';
         $allTags = Article::usedTagNames();
         $typeName = $this->categoryManager->typeName();
         $categories = $this->categoryManager->selectize();
-        $selectize_statuses = Status::of( Article::class )->groupByLevel();
+        $selectize_statuses = Status::of(Article::class)->groupByLevel();
         $this->buildHeading(
-            [ trans( 'common.create' ), $typeName ],
+            [trans('common.create'), $typeName],
             'plus-sign',
             [
-                route( 'backend.article.index' ) => $typeName,
-                '#'                              => trans( 'common.create' ),
+                route('backend.article.index') => $typeName,
+                '#' => trans('common.create'),
             ]
         );
 
-        return view( 'article::backend.form', compact( 'article', 'url', 'method', 'allTags', 'categories', 'selectize_statuses' ) );
+        return view('article::backend.form', compact('article', 'url', 'method', 'allTags', 'categories', 'selectize_statuses'));
     }
 
     /**
@@ -137,22 +148,23 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
-    public function store( ArticleRequest $request ) {
+    public function store(ArticleRequest $request)
+    {
         $article = new Article();
-        $article->fill( $request->all() );
-        $article->fillFeaturedImage( $request, true );
-        $article->user_id = user( 'id' );
+        $article->fill($request->all());
+        $article->fillFeaturedImage($request, true);
+        $article->user_id = user('id');
         //$article->fillStatus($request->get('s'));
         $article->save();
         Session::flash(
             'message',
             [
-                'type'    => 'success',
-                'content' => trans( 'common.create_object_success', [ 'name' => $this->categoryManager->typeName() ] ),
+                'type' => 'success',
+                'content' => trans('common.create_object_success', ['name' => $this->categoryManager->typeName()]),
             ]
         );
 
-        return redirect( route( 'backend.article.index' ) );
+        return redirect(route('backend.article.index'));
     }
 
     /**
@@ -160,25 +172,26 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\View\View
      */
-    public function show( Article $article ) {
+    public function show(Article $article)
+    {
         $typeName = $this->categoryManager->typeName();
         $this->buildHeading(
-            [ trans( 'common.view_detail' ), $typeName ],
+            [trans('common.view_detail'), $typeName],
             'list',
             [
-                route( 'backend.article.index' ) => $typeName,
-                '#'                              => trans( 'common.view_detail' ),
+                route('backend.article.index') => $typeName,
+                '#' => trans('common.view_detail'),
             ],
             [
                 [
-                    route( 'backend.article.edit', [ 'article' => $article->id ] ),
-                    trans( 'common.edit' ),
-                    [ 'type' => 'primary', 'size' => 'sm', 'icon' => 'edit' ],
+                    route('backend.article.edit', ['article' => $article->id]),
+                    trans('common.edit'),
+                    ['type' => 'primary', 'size' => 'sm', 'icon' => 'edit'],
                 ],
             ]
         );
 
-        return view( 'article::backend.show', compact( 'article', 'typeName' ) );
+        return view('article::backend.show', compact('article', 'typeName'));
     }
 
     /**
@@ -186,8 +199,9 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\View\View
      */
-    public function preview( Article $article ) {
-        return view( 'article::backend.preview', compact( 'article' ) );
+    public function preview(Article $article)
+    {
+        return view('article::backend.preview', compact('article'));
     }
 
     /**
@@ -195,23 +209,25 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\View\View
      */
-    public function edit( Article $article ) {
-        $url = route( 'backend.article.update', [ 'article' => $article->id ] );
+    public function edit(Article $article)
+    {
+        abort_unless($article->isReady('update'), 403, trans('common.forbidden'));
+        $url = route('backend.article.update', ['article' => $article->id]);
         $method = 'put';
         $allTags = Article::usedTagNames();
         $typeName = $this->categoryManager->typeName();
         $categories = $this->categoryManager->selectize();
-        $selectize_statuses = Status::of( Article::class )->groupByLevel();
+        $selectize_statuses = Status::of(Article::class)->groupByLevel();
         $this->buildHeading(
-            [ trans( 'common.update' ), $typeName ],
+            [trans('common.update'), $typeName],
             'edit',
             [
-                route( 'backend.article.index' ) => $typeName,
-                '#'                              => trans( 'common.edit' ),
+                route('backend.article.index') => $typeName,
+                '#' => trans('common.edit'),
             ]
         );
 
-        return view( 'article::backend.form', compact( 'article', 'categories', 'url', 'method', 'allTags', 'selectize_statuses' ) );
+        return view('article::backend.form', compact('article', 'categories', 'url', 'method', 'allTags', 'selectize_statuses'));
     }
 
     /**
@@ -220,20 +236,23 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\Http\RedirectResponse|\Illuminate\Routing\Redirector
      */
-    public function update( ArticleRequest $request, Article $article ) {
-        $article->fill( $request->all() );
-        $article->fillFeaturedImage( $request, true );
+    public function update(ArticleRequest $request, Article $article)
+    {
+        $article->timestamps = false; // TODO Set tạm thời
+        abort_unless($article->isReady('update'), 403, trans('common.forbidden'));
+        $article->fill($request->all());
+        $article->fillFeaturedImage($request, true);
         //$article->fillStatus($request->get('s'));
         $article->save();
         Session::flash(
             'message',
             [
-                'type'    => 'success',
-                'content' => trans( 'common.update_object_success', [ 'name' => $this->categoryManager->typeName() ] ),
+                'type' => 'success',
+                'content' => trans('common.update_object_success', ['name' => $this->categoryManager->typeName()]),
             ]
         );
 
-        return redirect( route( 'backend.article.index' ) );
+        return redirect(route('backend.article.index'));
     }
 
     /**
@@ -242,13 +261,15 @@ class BackendController extends BaseController {
      * @return \Illuminate\Http\JsonResponse
      * @throws \Exception
      */
-    public function destroy( Article $article ) {
+    public function destroy(Article $article)
+    {
+        abort_unless($article->isReady('delete'), 403, trans('common.forbidden'));
         $article->delete();
 
         return response()->json(
             [
-                'type'    => 'success',
-                'content' => trans( 'common.delete_object_success', [ 'name' => $this->categoryManager->typeName() ] ),
+                'type' => 'success',
+                'content' => trans('common.delete_object_success', ['name' => $this->categoryManager->typeName()]),
             ]
         );
     }
@@ -259,10 +280,12 @@ class BackendController extends BaseController {
      *
      * @return \Illuminate\Http\JsonResponse
      */
-    public function status( Article $article, $status ) {
-        $result = $article->update( [ 'status' => $status ] ) ? 'success' : 'error';
+    public function status(Article $article, $status)
+    {
+        abort_unless($article->isReady('update'), 403, trans('common.forbidden'));
+        $result = $article->update(['status' => $status]) ? 'success' : 'error';
 
-        return response()->json( [ 'type' => $result, 'content' => trans( "common.status_{$result}" ) ] );
+        return response()->json(['type' => $result, 'content' => trans("common.status_{$result}")]);
     }
 
     /**
@@ -270,11 +293,12 @@ class BackendController extends BaseController {
      *
      * @return array
      */
-    protected function quickUpdateAttributes() {
+    protected function quickUpdateAttributes()
+    {
         return [
             'title' => [
                 'rules' => 'required|max:255',
-                'label' => trans( 'article::common.title' ),
+                'label' => trans('article::common.title'),
             ],
         ];
     }
